Prevent history command from being recorded
If the assumed HISTCONTROL settings are not present, the history command is
written to the saved file, causing the history command to reappear when
restoring panes. This longer command prevents the saving of the command
regardless of the HISTCONTROL settings by writing it out and then copping
off the final history line containing the write command. The long command
also ends with clearing and restoring the local history with what was recorded
in order to prevent the history writing command from leaking into the saved
file when performing multiple tmux save actions.