Use 'cat' command always
`pane_creation_command` in `scripts/restore.sh` uses the string `cat` to
run `/bin/cat`; however, because `pane_creation_command` is evaluated in
the shell, aliases of `cat` may cause unexpected results; in my case, it
runs the `bat` [1] binary, which produces weird and garbled output. By
using `command -v`, we can get the path of the `cat` binary and avoid
this. (Note that we don't use `which` because `which` isn't
POSIX-standard.)

[1]: https://github.com/sharkdp/bat