Preserve the value of automatic-rename option
When windows are renamed while restoring tmux, the value of
automatic-rename is set to off, which might not be desired.

Therefore, the original value of the automatic-rename option for each
window is now saved and set again while restoring. This means that
windows intended to have fixed names will continue to have fixed names,
while windows intended to have automatic-rename will continue to have
automatic-rename.